Moved homepage and AppDrawer to seperate files
